Use reg key from installer for enabling updates
This will cause zipped versions to not be enabled for auto updates. There are many potential corner cases, e.g. there is an installed version, which mean the registry key exists, but if you run the zipped exe it will think it was installed. Other cases such as admin vs local installs can mess things up. Trying to handle every possible scenario is not worth it currently, so this is pretty simple as it checks to see if any of the appropriate locations indicate it is allowed to check for updates. Closes #43