Fix memory leak
Using ScintillaEdit::get_doc() caused a new ScintillaDocument pointer to get returned. A few places that used this call was for the lexer_changed() signal. Since it's difficult to properly destroy the ScintillaDocument that is returned from that call, this has ScintillaNext own a permanent pointer to it's ScintillaDocument and emits the lexerChanged signal through ScintillaNext. Closes #170